Implement a loading indicator while the theme is fetched from shared preferences
