+++ lib/app/home/widgets/custom_circular_progress_indicator.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class CustomCircularProgressIndicator extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        body: Center(
+            child: CircularProgressIndicator.adaptive()),
+      ),
+    );
+  }
+}
