+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:todo_tracker/app/about/about_screen.dart';
-import 'package:todo_tracker/app/home/home_screen.dart';
 import 'package:todo_tracker/app/home/home_view_model.dart';
+import 'package:todo_tracker/app/home/widgets/custom_circular_progress_indicator.dart';
+import 'package:todo_tracker/app/home/widgets/my_app_with_theme.dart';
 import 'package:todo_tracker/theme/theme_view_model.dart';
 
 void main() {
@@ -22,33 +22,15 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return Consumer<ThemeViewModel>(
       builder: (context, themeManager, _) => FutureBuilder<ThemeData>(
-        future: themeManager.getTheme(),
+        future: themeManager.theme,
         builder: (context, snapshot) {
           if (snapshot.hasData) {
             return MyAppWithTheme(snapshot.data!);
           } else {
-            return CircularProgressIndicator.adaptive();
+            return CustomCircularProgressIndicator();
           }
         },
       ),
-    );
-  }
-}
-
-class MyAppWithTheme extends StatelessWidget {
-  final ThemeData data;
-
-  MyAppWithTheme(this.data);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      theme: data,
-      routes: {
-        '/': (context) => HomeScreen(),
-        '/about': (context) => AboutPage(),
-      },
-      initialRoute: '/',
     );
   }
 }
